New features
Fixed "tom tom" incorrectly called "bass drum" in output messages.

Fixed multi-note pitchbends by removing portamento-to-note and adjusting
standard pitchbend range instead, thanks to a suggestion by Xky
(xkyrauh2001@hotmail.com)

Implemented a better method for reading Adlib register -> MIDI patch mapping
information (all stored in inst.txt now instead of having a seperate file for
each instrument.)  Also improved method for mapping instruments to percussion
on MIDI channel 10.

Fixed OPL rhythm instrument conversion issue (a MIDI noteon was being generated
too often - if the OPL instrument is on and we receive another keyon, it
*shouldn't* generate a fresh MIDI keyon.)

Fixed IMF type-1 conversion issue where unsigned numbers were being read as
signed, and the conversion was cutting off half way through.

+++ freq.cpp
@@ -13,7 +13,7 @@ double freq2key(int freq, int octave)
     return 0;*/
 	int iFNum = freq;
 	int iBlock = octave;
-	double dbOriginalFreq = 49716.0 * (double)iFNum * pow(2, iBlock - 20);
+	double dbOriginalFreq = 49716.0 * (double)iFNum * pow(2, (double)(iBlock - 20));
 	return 69.0 + 12.0 * log2(dbOriginalFreq / 440.0);
 }
 /*
